Change list name to fix typo error
I have used descriptive namings for my variables but I still missed
the fact the I'm comparing for each extension in JSON file with the
extensions in the file itself rather than the installed extensions
in the current machine.

I made this on my PC and when I ran this on my laptop, it removes
some of the items in the JSON list. I later found out even with new
and fresh install it detects as installed but still removes them
from the file.

Now it is working.